Add libtcod pseudo-terminal ASCII support
To use, define TCOD and provide a path to libtcod in your config.lua,
e.g.

    FPC_CMD = "fpc -dTCOD -Fl/path/to/libtcod"

`TTextIODriver.Create` will then have an extra boolean argument to
enable the libtcod-powered pseudo-terminal.

The biggest impact of using libtcod is that it doesn't require escape to
be pressed twice to register once, and more keys are recognized, such as
the arrow keys, the function keys and Ctrl and Alt.

You may need to fill in `LD_LIBRARY_PATH` when running programs compiled
with libtcod support on Linux, e.g. DoomRL needs to be built like this:

    LD_LIBRARY_PATH=/path/to/libtcod lua makefile.lua